requirements.txt
django-taggit
docutils
pygments
django_markup
django-markitup
